Fix comments from PR review
